Doing LTM searches when triggered by MemorySearchRequestEvent.
However: FindMatchingConceptProcessor expects a MemoryEvent. From LTM
I'm getting a PerceptEvent that wraps a Percept. In order to return it
as a MemoryEvent, I'm further wrapping that as a MemoryEvent...or
shifting the Percept to a MemoryEvent maybe?
But I'm also confused about what to do about returning multiple results
from memory in one go. So all up, need to think about this more.